fix a bug. daily commit.

+++ MIC_Calc_Function.h
@@ -22,6 +22,6 @@ __ONMIC__ void MIC_Node_Parallel(int n, int m, int *R, int *F, int *C,
 __ONMIC__ void MIC_WorkEfficient_Parallel(int n, int m, int *R, int *F, int *C,
 		float *result_mic);
 __ONMIC__ void MIC_Opt_BC(const int n, const int m, const int *R, const int *F,
-		const int *C, float *result_mic, const int num_cores);
+		const int *C, double *result_mic, const int num_cores);
 
 #endif /* MIC_CALC_FUNCTION_H_ */
